extension init needs to get called before open
