+++ sqlite3/lib/src/ffi/impl/database.dart
@@ -434,18 +434,4 @@ class DatabaseImpl extends Database {
       _finalizable._statements.remove(stmt._finalizable);
     }
   }
-
-  /// Load statically linked extension
-  void _autoLoadExtension(String extensionEntrypoint) {
-    final extensionPtr = _library.library.lookup<Void>(extensionEntrypoint);
-    final result = _bindings.sqlite3_auto_extension(extensionPtr);
-    if (result != SqlError.SQLITE_OK) {
-      throw SqliteException(result,
-          'Could not load extension with entrypoint "$extensionEntrypoint"');
-    }
-  }
-
-  @override
-  void ensureExtensionLoaded(LoadableExtension extension) =>
-      _autoLoadExtension(loadableExtensionEntrypoints[extension]!);
 }
